+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorSingleSchemaTest.kt
@@ -1,3 +1,28 @@
+/*
+ * @(#) CodeGeneratorSingleSchemaTest.kt
+ *
+ * json-kotlin-schema-codegen  JSON Schema Code Generation
+ * Copyright (c) 2020 Peter Wall
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package net.pwall.json.schema.codegen
 
 import kotlin.test.Test
@@ -7,6 +32,8 @@ import java.io.File
 import java.io.StringWriter
 
 import net.pwall.json.schema.JSONSchema
+import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.createHeader
+import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.outputCapture
 
 class CodeGeneratorSingleSchemaTest {
 
@@ -16,11 +43,10 @@ class CodeGeneratorSingleSchemaTest {
         val codeGenerator = CodeGenerator()
         codeGenerator.baseDirectoryName = "dummy"
         val stringWriter = StringWriter()
-        codeGenerator.outputResolver =
-                CodeGeneratorTestUtil.outputCapture("dummy", emptyList(), "Test", "kt", stringWriter)
+        codeGenerator.outputResolver = outputCapture("dummy", emptyList(), "Test", "kt", stringWriter)
         codeGenerator.basePackageName = "com.example"
         codeGenerator.generateClass(schema, "Test")
-        expect(CodeGeneratorExampleTest.expectedExample) { stringWriter.toString() }
+        expect(createHeader("Test") + CodeGeneratorExampleTest.expectedExample) { stringWriter.toString() }
     }
 
     @Test fun `should generate from pre-loaded schema as array`() {
@@ -29,11 +55,10 @@ class CodeGeneratorSingleSchemaTest {
         val codeGenerator = CodeGenerator()
         codeGenerator.baseDirectoryName = "dummy"
         val stringWriter = StringWriter()
-        codeGenerator.outputResolver =
-                CodeGeneratorTestUtil.outputCapture("dummy", emptyList(), "Test", "kt", stringWriter)
+        codeGenerator.outputResolver = outputCapture("dummy", emptyList(), "Test", "kt", stringWriter)
         codeGenerator.basePackageName = "com.example"
         codeGenerator.generateClasses(listOf(schema to "Test"))
-        expect(CodeGeneratorExampleTest.expectedExample) { stringWriter.toString() }
+        expect(createHeader("Test") + CodeGeneratorExampleTest.expectedExample) { stringWriter.toString() }
     }
 
 }